# src/main/java/com/project/ecommerce/perfume/dto/user/BaseUserResponse.java
package com.project.ecommerce.perfume.dto.user;

import com.project.ecommerce.perfume.enums.Role;
import lombok.Getter;
import lombok.Setter;

import java.util.Set;

@Getter
@Setter
public class BaseUserResponse {
    private Long id;
    private String email;
    private String firstName;
    private Set<Role> roles;
    private String provider;
}